refactor: make pydcmqi robust and usable
- Target Python 3.10+ (drop 3.8/3.9), update CI matrix, mypy, pylint
- Split segimage.py (807 lines) into triplet.py, segment.py, segimage.py
- Expose full public API via __init__.py (SegImage, Segment, Triplet, etc.)
- Add DcmqiError exception with subprocess stderr capture
- Add shutil.which check before calling dcmqi CLI tools
- Expose dcmqi CLI options: merge_segments, skip_empty_slices, geometry_check
- Add Triplet.from_code()/to_code() for pydicom/highdicom interop
- Replace 5 assert statements with proper ValueError/TypeError/RuntimeError
- Fix load() return type annotation (-> bool to -> None)
- Fix parameter typo: diable_sanity_check -> disable_sanity_check
- Fix TypedDicts to match dcmqi seg-schema.json (required/optional split)
- Add SegmentAlgorithmType Literal and missing schema fields
- Add stdlib logging, replace commented-out print statements
- Fix broken markdown links in README, add quickstart and interop docs
- Add test_public_api and TestErrorPaths with 5 new error path tests
