+++ .github/workflows/ci.yml
@@ -42,12 +42,8 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: ["3.8", "3.12"]
+        python-version: ["3.10", "3.12"]
         runs-on: [ubuntu-latest, macos-latest, windows-latest]
-
-        include:
-          - python-version: pypy-3.10
-            runs-on: ubuntu-latest
 
     steps:
       - uses: actions/checkout@v4
