+++ noxfile.py
@@ -1,5 +1,3 @@
-from __future__ import annotations
-
 import argparse
 import shutil
 from pathlib import Path
