+++ tests/test_package.py
@@ -1,5 +1,3 @@
-from __future__ import annotations
-
 import importlib.metadata
 
 import pydcmqi as m
@@ -7,3 +5,24 @@ import pydcmqi as m
 
 def test_version():
     assert importlib.metadata.version("pydcmqi") == m.__version__
+
+
+def test_public_api():
+    """Verify all expected symbols are importable from the top-level package."""
+    from pydcmqi import (
+        DcmqiError,
+        SegImage,
+        SegImageData,
+        SegImageFiles,
+        Segment,
+        SegmentData,
+        Triplet,
+        SegImageDict,
+        SegmentDict,
+        TripletDict,
+    )
+
+    # Verify they are the right types
+    assert callable(SegImage)
+    assert callable(Triplet)
+    assert callable(DcmqiError)
